pretty safe and working version of a function that generates an emiter / listener url pair
